The change to given code was preventing saves.
I decided to upload my corrected code for the HW after our Student Success Manager spoke to class about the proper way to reference other's codes and no cloning. I was able to finish this assignment with the help of classmates Shannon Ervin and Sara Damyan as well as the TAs. On our study group, that also includes Marysia, we try to help each other as much as possible when there are questions and pointing to resources, but also in seeing the TAs. OH are limited, so whatever is important that 1 of us learns from it, we pass along to others.